system-design-challenges-50/day21/requirements.txt
fastapi==0.68.0
uvicorn[standard]==0.15.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
redis==3.5.3
celery==5.1.2
pydantic==1.8.2
pydantic-settings==2.0.3
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.5
Pillow==8.3.1
alembic==1.7.1
pytest==6.2.4
pytest-asyncio==0.15.1
httpx==0.18.2